feat: split dns and dnssec stacks
