+++ src/ApiStage.ts
@@ -1,6 +1,9 @@
 import { Stage, StageProps } from 'aws-cdk-lib';
 import { Construct } from 'constructs';
 import { ApiStack } from './ApiStack';
+import { DnsSecStack } from './DnsSecStack';
+import { DnsStack } from './DnsStack';
+import { Statics } from './Statics';
 
 export interface ApiStageProps extends StageProps {
   branch: string;
@@ -17,6 +20,17 @@ export class ApiStage extends Stage {
     new ApiStack(this, 'api-stack', {
       branch: props.branch,
     });
+
+    new DnsStack(this, 'dns-stack', {
+      branch: props.branch,
+    });
+
+    // Only deploy dnssec on non dev branches
+    if (!Statics.isDevelopment(props.branch)) {
+      new DnsSecStack(this, 'dnssec-stack', {
+        branch: props.branch,
+      });
+    }
 
     // cfn-nag for all stacks in this stage.
     //Aspects.of(apiStack).add(new AwsSolutionsChecks());
